refact: performance

+++ internal/payments/entities/summary.go
@@ -1,7 +1,7 @@
 package entities
 
 type Summary struct {
-	TotalRequests int64   `json:"totalRequests"`
+	TotalRequests int     `json:"totalRequests"`
 	TotalAmount   float64 `json:"totalAmount"`
 }
 
@@ -11,6 +11,6 @@ type AggregatedSummary struct {
 }
 
 func (s *AggregatedSummary) RoundAmount() {
-	s.Default.TotalAmount = float64(int64(s.Default.TotalAmount*100+0.5)) / 100
-	s.Fallback.TotalAmount = float64(int64(s.Fallback.TotalAmount*100+0.5)) / 100
+	s.Default.TotalAmount = float64(int(s.Default.TotalAmount*100+0.5)) / 100
+	s.Fallback.TotalAmount = float64(int(s.Fallback.TotalAmount*100+0.5)) / 100
 }
